add setup
